Use uint64_t instead of long long unsigned for a 64-bit unsigned integer.  Update the ddcMD distribution with new version of object.c that supports this usage.  Note that uint64_t is a different type on different platforms, but is always 64 bits.
git-svn-id: svn+ssh://lc-svn.llnl.gov/usr/gapps/emhm/svn/Cardioid/EP/trunk@998 baecf98e-1e81-4516-8920-d4c889e67ea4

+++ src/Long64.hh
@@ -1,6 +1,7 @@
 #ifndef LONG64_H
 #define LONG64_H
 
-typedef unsigned long long Long64;
+#include <stdint.h>
+typedef uint64_t Long64;
 
 #endif
